Renamed GeneralContent to Content and updated Tests
