+++ src/Content/ContentTitle.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const ContentTitle = ({ children }) =>
+  <h2>
+    {children}
+  </h2>
+
+const defaultProps = {};
+const propTypes = {};
+
+ContentTitle.defaultProps = defaultProps;
+ContentTitle.propTypes = propTypes;
+
+export default ContentTitle;
