--- src/GeneralContent/ContentTitle.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const GeneralContentTitle = ({ children }) =>
-  <h2>
-    {children}
-  </h2>
-
-const defaultProps = {};
-const propTypes = {};
-
-GeneralContentTitle.defaultProps = defaultProps;
-GeneralContentTitle.propTypes = propTypes;
-
-export default GeneralContentTitle;
