+++ src/index.js
@@ -1,7 +1,6 @@
 import * as Accordion from './Accordion'
-import * as CallsToAction from './CallsToAction'
-import * as GeneralContent from './GeneralContent'
+import * as Content from './Content'
 import * as Loaders from './Loaders'
 import * as Media from './Media'
 
-export { Accordion, CallsToAction, GeneralContent, Loaders, Media }
+export { Accordion, Content, Loaders, Media }
